+++ src/utils/color/hsb.rs
@@ -2,14 +2,14 @@ use std::convert::From;
 use super::RgbColor;
 
 pub struct HsbColor {
-    pub hue : f64,
-    pub saturation : f64,
-    pub brightness : f64,
+    pub hue : f32,
+    pub saturation : f32,
+    pub brightness : f32,
 }
 
 impl HsbColor {
 
-    pub fn new(hue : f64, saturation : f64, brightness : f64) -> Self {
+    pub fn new(hue : f32, saturation : f32, brightness : f32) -> Self {
         //TODO : treat invalid parameters case 
         Self {
             hue,
@@ -21,18 +21,18 @@ impl HsbColor {
 
 impl From<RgbColor> for HsbColor  {
     fn from(color : RgbColor) -> Self {
-        let r = color.red as f64 / 255f64;
-        let g = color.green as f64 / 255f64;
-        let b = color.blue as f64 / 255f64;
+        let r = color.red as f32 / 255f32;
+        let g = color.green as f32 / 255f32;
+        let b = color.blue as f32 / 255f32;
         let max = r.max(g.max(b));
         let min = r.min(g.min(b));
         let range = max - min;
         let hue = match (range, max == r, max == g, max == b) {
-            (0f64,_,_,_)    => 0f64,
-            (_,true,_,_) => ((g - b) / range % 6f64 ) * 60f64,
-            (_,_,true,_) => ((b - r) / range + 2f64) * 60f64,
-            (_,_,_,true) => ((r - g) / range + 4f64) * 60f64,
-            (_, false, false, false) => 0f64,
+            (0f32,_,_,_)    => 0f32,
+            (_,true,_,_) => ((g - b) / range % 6f32 ) * 60f32,
+            (_,_,true,_) => ((b - r) / range + 2f32) * 60f32,
+            (_,_,_,true) => ((r - g) / range + 4f32) * 60f32,
+            (_, false, false, false) => 0f32,
         };
 
         let brightness = max;
