+++ src/utils/color/rgb.rs
@@ -17,8 +17,8 @@ impl RgbColor {
         }
     }
 
-    pub fn to_vec(&self) -> Vec<f64> {
-        return vec![self.red as f64, self.green as f64, self.blue as f64];
+    pub fn to_vec(&self) -> Vec<f32> {
+        return vec![self.red as f32, self.green as f32, self.blue as f32];
     }
 }
 
@@ -29,21 +29,21 @@ impl From<HsbColor> for RgbColor {
         let mean = color.brightness - range;
 
         let (rp, gp, bp) = {
-            let x = range * (1f64 - (color.hue / 60f64 % 2f64 - 1f64).abs());
+            let x = range * (1f32 - (color.hue / 60f32 % 2f32 - 1f32).abs());
             match color.hue as u16 / 60u16 % 6  {
-                0 => (range, x, 0f64),
-                1  => (x, range, 0f64),
-                2 => (0f64, range, x),
-                3 => (0f64, x, range),
-                4 => (x, 0f64, range),
-                5 => (range, 0f64, x),
-                _ => (0f64, 0f64, 0f64),
+                0 => (range, x, 0f32),
+                1  => (x, range, 0f32),
+                2 => (0f32, range, x),
+                3 => (0f32, x, range),
+                4 => (x, 0f32, range),
+                5 => (range, 0f32, x),
+                _ => (0f32, 0f32, 0f32),
             }
         };
         Self {
-            red : ((rp + mean) * 255f64) as u8,
-            green : ((gp + mean) * 255f64) as u8,
-            blue : ((bp + mean) * 255f64) as u8,
+            red : ((rp + mean) * 255f32) as u8,
+            green : ((gp + mean) * 255f32) as u8,
+            blue : ((bp + mean) * 255f32) as u8,
         }
     }
 }
